--- src/Module/Application/Admin/Filter/BrowseFiltersAction.php
@@ -1,63 +1,0 @@
-<?php
-/*
- * vim:set softtabstop=4 shiftwidth=4 expandtab:
- *
- * LICENSE: GNU Affero General Public License, version 3 (AGPL-3.0-or-later)
- * Copyright 2001 - 2020 Ampache.org
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *
- */
-
-declare(strict_types=0);
-
-namespace Ampache\Module\Application\Admin\Filter;
-
-use Ampache\Module\Application\ApplicationActionInterface;
-use Ampache\Module\Application\Exception\AccessDeniedException;
-use Ampache\Module\Authorization\AccessLevelEnum;
-use Ampache\Module\Authorization\GuiGatekeeperInterface;
-use Ampache\Module\Util\Ui;
-use Ampache\Module\Util\UiInterface;
-use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\ServerRequestInterface;
-
-final class BrowseFiltersAction implements ApplicationActionInterface
-{
-    public const REQUEST_KEY = 'browse_catalog_filters';
-
-    private UiInterface $ui;
-
-    public function __construct(
-        UiInterface $ui
-    ) {
-        $this->ui = $ui;
-    }
-
-    public function run(ServerRequestInterface $request, GuiGatekeeperInterface $gatekeeper): ?ResponseInterface
-    {
-        if ($gatekeeper->mayAccess(AccessLevelEnum::TYPE_INTERFACE, AccessLevelEnum::LEVEL_MANAGER) === false) {
-            throw new AccessDeniedException();
-        }
-
-        $this->ui->showHeader();
-
-        require_once Ui::find_template('show_manage_filters.inc.php');
-
-        $this->ui->showQueryStats();
-        //$this->ui->showFooter();
-
-        return null;
-    }
-}
