+++ src/Module/Application/Admin/Filter/ConfirmDeleteAction.php
@@ -26,9 +26,6 @@ namespace Ampache\Module\Application\Admin\Filter;
 
 use Ampache\Config\ConfigContainerInterface;
 use Ampache\Config\ConfigurationKeyEnum;
-use Ampache\Repository\Model\ModelFactoryInterface;
-use Ampache\Module\Application\Exception\AccessDeniedException;
-use Ampache\Module\System\Core;
 use Ampache\Module\Util\UiInterface;
 use Ampache\Repository\Model\Catalog;
 use Psr\Http\Message\ResponseInterface;
@@ -70,13 +67,13 @@ final class ConfirmDeleteAction extends AbstractFilterAction
             $this->ui->showConfirmation(
                 T_('No Problem'),
                 sprintf(T_('%s has been deleted'), $filter_name),
-                sprintf('%s/admin/filter.php?action=browse_catalog_filters', $this->configContainer->getWebPath())
+                sprintf('%s/admin/filter.php?action=show', $this->configContainer->getWebPath())
             );
         } else {
             $this->ui->showConfirmation(
                 T_('There Was a Problem'),
                 T_('You need at least one active Administrator account'),
-                sprintf('%s/admin/filter.php?action=browse_catalog_filters', $this->configContainer->getWebPath())
+                sprintf('%s/admin/filter.php?action=show', $this->configContainer->getWebPath())
             );
         }
